+++ client/src/main/kotlin/org/dreamexposure/discal/client/config/DiscordConfig.kt
@@ -1,9 +1,5 @@
 package org.dreamexposure.discal.client.config
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import com.fasterxml.jackson.module.kotlin.registerKotlinModule
-import discord4j.common.JacksonResources
 import discord4j.common.store.Store
 import discord4j.common.store.legacy.LegacyStoreLayout
 import discord4j.core.DiscordClientBuilder
@@ -31,24 +27,12 @@ import kotlinx.coroutines.reactor.mono
 import org.dreamexposure.discal.Application
 import org.dreamexposure.discal.client.listeners.discord.EventListener
 import org.dreamexposure.discal.core.config.Config
-import org.dreamexposure.discal.core.serializers.SnowflakeMapper
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Primary
 import reactor.kotlin.core.publisher.toFlux
 
 @Configuration
 class DiscordConfig {
-
-    @Bean
-    @Primary
-    fun objectMapper(): ObjectMapper {
-        // Use d4j's object mapper
-        return JacksonResources.create().objectMapper
-            .registerKotlinModule()
-            .registerModule(JavaTimeModule())
-            .registerModule(SnowflakeMapper())
-    }
 
     @Bean
     fun discordGatewayClient(
