+++ client/src/main/kotlin/org/dreamexposure/discal/client/listeners/discord/GuildCreateListener.kt
@@ -1,7 +1,7 @@
 package org.dreamexposure.discal.client.listeners.discord
 
+import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
-import discord4j.common.JacksonResources
 import discord4j.core.event.domain.guild.GuildCreateEvent
 import discord4j.discordjson.json.ApplicationCommandRequest
 import kotlinx.coroutines.reactor.awaitSingle
@@ -13,18 +13,17 @@ import org.springframework.core.io.support.PathMatchingResourcePatternResolver
 import org.springframework.stereotype.Component
 
 @Component
-class GuildCreateListener : EventListener<GuildCreateEvent> {
+class GuildCreateListener(objectMapper: ObjectMapper) : EventListener<GuildCreateEvent> {
     private val premiumCommands: List<ApplicationCommandRequest>
     private val devCommands: List<ApplicationCommandRequest>
 
     init {
-        val d4jMapper = JacksonResources.create()
         val matcher = PathMatchingResourcePatternResolver()
 
         // Get premium commands
         val premiumCommands = mutableListOf<ApplicationCommandRequest>()
         for (res in matcher.getResources("commands/premium/*.json")) {
-            val request = d4jMapper.objectMapper.readValue<ApplicationCommandRequest>(res.inputStream)
+            val request = objectMapper.readValue<ApplicationCommandRequest>(res.inputStream)
             premiumCommands.add(request)
         }
         this.premiumCommands = premiumCommands
@@ -32,7 +31,7 @@ class GuildCreateListener : EventListener<GuildCreateEvent> {
         // Get dev commands
         val devCommands = mutableListOf<ApplicationCommandRequest>()
         for (res in matcher.getResources("commands/dev/*.json")) {
-            val request = d4jMapper.objectMapper.readValue<ApplicationCommandRequest>(res.inputStream)
+            val request = objectMapper.readValue<ApplicationCommandRequest>(res.inputStream)
             premiumCommands.add(request)
         }
         this.devCommands = devCommands
