+++ core/src/main/kotlin/org/dreamexposure/discal/core/config/BeanConfig.kt
@@ -1,11 +1,26 @@
 package org.dreamexposure.discal.core.config
 
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
+import com.fasterxml.jackson.module.kotlin.registerKotlinModule
+import discord4j.common.JacksonResources
+import org.dreamexposure.discal.core.serializers.SnowflakeMapper
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Primary
 import org.springframework.web.reactive.result.method.annotation.RequestMappingHandlerMapping
 
 @Configuration
 class BeanConfig {
+    @Bean
+    @Primary
+    fun objectMapper(): ObjectMapper {
+        // Use d4j's object mapper
+        return JacksonResources.create().objectMapper
+            .registerKotlinModule()
+            .registerModule(JavaTimeModule())
+            .registerModule(SnowflakeMapper())
+    }
 
     @Bean
     fun handlerMapping(): RequestMappingHandlerMapping {
