+++ server/src/main/kotlin/org/dreamexposure/discal/server/network/discord/GlobalCommandRegistrar.kt
@@ -1,7 +1,7 @@
 package org.dreamexposure.discal.server.network.discord
 
+import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
-import discord4j.common.JacksonResources
 import discord4j.discordjson.json.ApplicationCommandRequest
 import discord4j.rest.RestClient
 import org.dreamexposure.discal.core.logger.LOGGER
@@ -13,19 +13,18 @@ import org.springframework.stereotype.Component
 
 @Component
 class GlobalCommandRegistrar(
-        private val restClient: RestClient
+    private val restClient: RestClient,
+    private val objectMapper: ObjectMapper,
 ) : ApplicationRunner {
 
     override fun run(args: ApplicationArguments?) {
-        val d4jMapper = JacksonResources.create()
-
         val matcher = PathMatchingResourcePatternResolver()
         val applicationService = restClient.applicationService
         val applicationId = restClient.applicationId.block()!!
 
         val commands = mutableListOf<ApplicationCommandRequest>()
         for (res in matcher.getResources("commands/global/*.json")) {
-            val request = d4jMapper.objectMapper.readValue<ApplicationCommandRequest>(res.inputStream)
+            val request = objectMapper.readValue<ApplicationCommandRequest>(res.inputStream)
             commands.add(request)
         }
